FIX: replace build only if needed

+++ env/bsc.mn5.gcc.cuda.mkl.ss.sh
@@ -20,8 +20,12 @@ fi
 
 buildname="build-bsc-mn5-gcc-cuda${cudaversionname}-mkl-ss"
 export WAFLOCK=".lock-waf_linux_${buildname}"
-rm -rf build
-ln -s "${buildname}" build
+curr_buildname="$(readlink build)"
+if [ "${curr_buildname}" != "${buildname}" ]
+then
+    rm -rf build
+    ln -s "${buildname}" build
+fi
