+++ env/it4i.cs.spr.sh
@@ -18,8 +18,12 @@ fi
 
 buildname="build-it4i-cs-spr-${cpu_spsolver}"
 export WAFLOCK=".lock-waf_linux_${buildname}"
-rm -rf build
-ln -s "${buildname}" build
+curr_buildname="$(readlink build)"
+if [ "${curr_buildname}" != "${buildname}" ]
+then
+    rm -rf build
+    ln -s "${buildname}" build
+fi
