+++ env/it4i.karolina.intel.cuda.mkl.ss.sh
@@ -20,8 +20,12 @@ fi
 
 buildname="build-it4i-karolina-intel-cuda-mkl-ss"
 export WAFLOCK=".lock-waf_linux_${buildname}"
-rm -rf build
-ln -s "${buildname}" build
+curr_buildname="$(readlink build)"
+if [ "${curr_buildname}" != "${buildname}" ]
+then
+    rm -rf build
+    ln -s "${buildname}" build
+fi
